requirements.txt
PySide6>=6.7.0
opencv-python>=4.9.0
Pillow>=10.0.0
numpy>=1.26.0
